Code Style and combine two test cases
