+++ test/java/seedu/addressbook/util/TypicalPersons.java
@@ -19,56 +19,55 @@ import seedu.addressbook.data.tag.UniqueTagList;
  */
 public class TypicalPersons {
 
-     public Person amy, bill, candy, dan;
+    public Person amy, bill, candy, dan;
 
-     public TypicalPersons() {
-         try {
-             amy = new Person(new Name("Amy Buck"), new Phone("91119111", false), new Email("ab@gmail.com", false),
-                     new Address("1 Clementi Road", false), new UniqueTagList());
-             bill = new Person(new Name("Bill Clint"), new Phone("92229222", false), new Email("bc@gmail.com", false),
-                     new Address("2 Clementi Road", true), new UniqueTagList());
-             candy = new Person(new Name("Candy Destiny"), new Phone("93339333", true),
-                     new Email("cd@gmail.com", false), new Address("3 Clementi Road", true), new UniqueTagList());
-             dan = new Person(new Name("Dan Smith"), new Phone("1234556", true), new Email("ss@tt.com", true),
-                     new Address("NUS", true), new UniqueTagList(new Tag("Test")));
-         } catch (IllegalValueException e) {
-             e.printStackTrace();
-             assert false : "not possible";
-         }
-     }
+    public TypicalPersons() {
+        try {
+            amy = new Person(new Name("Amy Buck"), new Phone("91119111", false), new Email("ab@gmail.com", false),
+                    new Address("1 Clementi Road", false), new UniqueTagList());
+            bill = new Person(new Name("Bill Clint"), new Phone("92229222", false), new Email("bc@gmail.com", false),
+                    new Address("2 Clementi Road", true), new UniqueTagList());
+            candy = new Person(new Name("Candy Destiny"), new Phone("93339333", true),
+                    new Email("cd@gmail.com", false), new Address("3 Clementi Road", true), new UniqueTagList());
+            dan = new Person(new Name("Dan Smith"), new Phone("1234556", true), new Email("ss@tt.com", true),
+                    new Address("NUS", true), new UniqueTagList(new Tag("Test")));
+        } catch (IllegalValueException e) {
+            e.printStackTrace();
+            assert false : "not possible";
+        }
+    }
 
-     private void loadAddressBookWithSampleData(AddressBook ab) {
-         try {
-             for (Person p : this.getTypicalPersons()) {
-                 ab.addPerson(new Person(p));
-             }
-         } catch (IllegalValueException e) {
-             assert false : "not possible";
-         }
-     }
+    private void loadAddressBookWithSampleData(AddressBook ab) {
+        try {
+            for (Person p : this.getTypicalPersons()) {
+                ab.addPerson(new Person(p));
+            }
+        } catch (IllegalValueException e) {
+            assert false : "not possible";
+        }
+    }
 
-     public Person[] getTypicalPersons() {
-         return new Person[]{amy, bill, candy, dan};
-     }
+    public Person[] getTypicalPersons() {
+        return new Person[]{amy, bill, candy, dan};
+    }
 
-     public AddressBook getTypicalAddressBook() {
-         AddressBook ab = new AddressBook();
-         loadAddressBookWithSampleData(ab);
-         return ab;
-     }
+    public AddressBook getTypicalAddressBook() {
+        AddressBook ab = new AddressBook();
+        loadAddressBookWithSampleData(ab);
+        return ab;
+    }
 
-     public List<ReadOnlyPerson> getListWithAllPersons() {
-         List<ReadOnlyPerson> list = new ArrayList<ReadOnlyPerson>();
-         for (Person p : getTypicalPersons()) {
-             list.add(new Person(p));
-         }
-         return list;
-     }
+    public List<ReadOnlyPerson> getListWithAllPersons() {
+        List<ReadOnlyPerson> list = new ArrayList<ReadOnlyPerson>();
+        for (Person p : getTypicalPersons()) {
+            list.add(new Person(p));
+        }
+        return list;
+    }
 
-     public List<ReadOnlyPerson> getListWithSomePersons() {
-         List<ReadOnlyPerson> list = getListWithAllPersons();
-         list.remove(0); //remove a person in the list
-         return list;
-     }
- }
- 
+    public List<ReadOnlyPerson> getListWithSomePersons() {
+        List<ReadOnlyPerson> list = getListWithAllPersons();
+        list.remove(0); //remove a person in the list
+        return list;
+    }
+}
